requirements.txt
discord.py
pytz
python-dotenv
pymongo
dnspython